refactor: ATLAS-400: Restructure some existing MPA directories
