+++ Atlas.MatchPrediction.Functions/Functions/MatchCalculationFunctions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Threading.Tasks;
-using Atlas.MatchPrediction.ExternalInterface.Models.MatchCalculation;
+using Atlas.MatchPrediction.ExternalInterface.Models.MatchPredictionSteps.MatchCalculation;
 using Atlas.MatchPrediction.Services.MatchCalculation;
 using AzureFunctions.Extensions.Swashbuckle.Attribute;
 using Microsoft.AspNetCore.Http;
