+++ Atlas.MatchPrediction.Test/Services/HaplotypeFrequencies/FrequencyCsvReaderTests.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Linq;
 using System.Text;
-using Atlas.MatchPrediction.Services.HaplotypeFrequencies;
+using Atlas.MatchPrediction.Services.HaplotypeFrequencies.Import;
 using CsvHelper;
 using FluentAssertions;
 using NUnit.Framework;
