+++ Atlas.MatchPrediction/ExternalInterface/Models/MatchPredictionSteps/ExpandAmbiguousPhenotype/ExpandAmbiguousPhenotypeInput.cs
@@ -1,6 +1,6 @@
 ﻿using Atlas.Common.GeneticData.PhenotypeInfo;
 
-namespace Atlas.MatchPrediction.ExternalInterface.Models.ExpandAmbiguousPhenotype
+namespace Atlas.MatchPrediction.ExternalInterface.Models.MatchPredictionSteps.ExpandAmbiguousPhenotype
 {
     public class ExpandAmbiguousPhenotypeInput
     {
