+++ Atlas.MatchPrediction/ExternalInterface/Models/MatchPredictionSteps/ExpandAmbiguousPhenotype/ExpandAmbiguousPhenotypeResponse.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Atlas.Common.GeneticData.PhenotypeInfo;
 
-namespace Atlas.MatchPrediction.ExternalInterface.Models.ExpandAmbiguousPhenotype
+namespace Atlas.MatchPrediction.ExternalInterface.Models.MatchPredictionSteps.ExpandAmbiguousPhenotype
 {
     public class ExpandAmbiguousPhenotypeResponse
     {
