+++ Atlas.MatchPrediction/ExternalInterface/Models/MatchPredictionSteps/GenotypeLikelihood/GenotypeLikelihoodInput.cs
@@ -1,6 +1,6 @@
 ﻿using Atlas.Common.GeneticData.PhenotypeInfo;
 
-namespace Atlas.MatchPrediction.ExternalInterface.Models.GenotypeLikelihood
+namespace Atlas.MatchPrediction.ExternalInterface.Models.MatchPredictionSteps.GenotypeLikelihood
 {
     public class GenotypeLikelihoodInput
     {
